refactor: extract shared E2E helpers to tests/e2e/helpers.ts
Move generic helpers that can be reused by other E2E tests:
- TEST_SERVER_URL, RULES_TABLE_TIMEOUT, DIALOG_TIMEOUT constants
- setupConsoleErrorMonitoring(), assertNoConsoleErrors()
- clearAllRules(), reloadAndWaitForTable(), saveRule()

Toggle-specific helpers remain in the feature directory:
- TOGGLE_STATE_TIMEOUT, waitForToggleState(), getToggleState(), clickToggle()
